Se movio el metodo crear_datos_relevamiento a utils.py dentro de test
